Struct include_dir::IncludeDirBuilder

pub struct IncludeDirBuilder { /* fields omitted */ }

A builder object to give the include_dir() function a nice fluent API.

Note

It is pretty much assumed that you'll use it in the order described in include_dir's docs. Straying from that path will only lead to frustration.

Methods

impl IncludeDirBuilder

pub fn as_variable(self, name: &str) -> Self

Set the variable name to save your assets under.

pub fn ignore<S: AsRef<str>>(self, name: S) -> Self

Ignore a directory or file.

Note

For now, this will only match the path relative to the included directory.

pub fn to_file<P: AsRef<Path>>(self, path: P) -> Result<()>

Save the file tree to the provided path, performing all necessary error checking.